added db methods for displaying avail lockers to book

+++ app/src/main/java/com/example/lockerxlogin/Locker.java
@@ -2,14 +2,14 @@ package com.example.lockerxlogin;
 import java.util.ArrayList;
 
 public class Locker extends LockerStructure {
-    private int lockerID;
+    private long lockerID;
     private char size;  //'S'=small, 'M'=medium, 'L'=large
     private boolean isLocked; // False = Unlocked, True = Locked
     //private ArrayList<Booking> booking;
 
     public Locker(){} //empty constructor
 
-    public Locker(int structureID, int locationID, String location, String postalCode, String address, int lockerID, char size){
+    public Locker(long structureID, int locationID, String location, String postalCode, String address, int lockerID, char size){
         super.structureID = structureID;
 //        super.locationID = locationID;
 //        super.location = location;
@@ -21,10 +21,10 @@ public class Locker extends LockerStructure {
         this.isLocked = false; //unlocked
     }
 
-    public void setLockerID(int lockerID){
+    public void setLockerID(long lockerID){
         this.lockerID = lockerID;
     }
-    public int getLockerID(){
+    public long getLockerID(){
         return this.lockerID;
     }
